Make success child view of Estimate, rename pages dir
