+++ src/views/Privacy.vue
@@ -19,7 +19,7 @@
 
 <script>
 import VueMarkdown from 'vue-markdown'
-import text from 'raw-loader!./privacy-policy.md' //eslint-disable-line
+import text from 'raw-loader!../content/privacy-policy.md' //eslint-disable-line
 
 import HeroSection from '@/components/organisms/HeroSection'
 import MainNav from '@/components/organisms/MainNav'
